+++ src/Entities/Units/Units/Mage.java
@@ -11,7 +11,7 @@ import org.jetbrains.annotations.NotNull;
 import static Utilities.Constants.Colors.ANSI_RESET;
 
 public class Mage extends Unit implements IMage {
-    public Mage(String name, int hp, IDamage damage, int attackRange, int defence, float energy, int cost, String symbol, int row, int col, Player player) {
+    public Mage(String name, int hp, IDamage damage, int attackRange, int defence, double energy, int cost, String symbol, int row, int col, Player player) {
         super(name, hp, damage, attackRange, defence, energy, cost, symbol, row, col, player);
         getTerrains().put("*", 1F);
         getTerrains().put("#", 1.5F);
